requirements.txt
Flask
gunicorn
azure-storage-blob
azure-identity
pymysql